Move QueueWorker to its own module
